implements POST /api/categories functionality, and refactors test suite to test each table endpoints
