+++ errors/handlers.js
@@ -20,7 +20,8 @@ exports.handle400 = (error, req, res, next) => {
     || error.code === "2201X" 
     || error.code === "22P02"
     || error.code === "23502"
-    || error.code === "42703") {
+    || error.code === "42703"
+    || error.code === "23505") {
       
         res.status(400).send({msg: "bad request"})
     } else if (error.code === "23503") {
