+++ models/categories.models.js
@@ -6,3 +6,14 @@ exports.selectCategories = () => {
         return categories
     })
 }
+
+exports.insertCategory = (newCategory) => {
+    const categoryData = [newCategory.slug, newCategory.description]
+    return db.query("INSERT INTO categories (slug, description) VALUES ($1, $2) RETURNING *", categoryData).then(({rows : [insertedCategory]}) => {
+        if (!insertedCategory) {
+            return Promise.reject({status: 404, msg: "not found"})
+        } else {
+            return insertedCategory
+        }
+    })
+}
