+++ package.json
@@ -8,7 +8,7 @@
     "seed-prod": "NODE_ENV=production npm run seed",
     "setup-dbs": "psql -f ./db/setup.sql",
     "seed": "node ./db/seeds/run-seed.js",
-    "test": "jest",
+    "test": "jest --runInBand",
     "prepare": "husky install",
     "dev": "node ./db/seeds/run-seed.js && nodemon listen.js"
   },
